Create README.md
Welcome to the repository for my Google data analytics capstone project! This project was completed as part of the Google data analytics course offered by Coursera. The goal of the project was to apply the skills and knowledge learned throughout the course to a real-world data analytics problem.

In this project, I was required to help Cyclistic, a bikeshare company understand how casual and annual members differ. The results from the analysis would be used by the management in designing marketing strategies to get more casuals to acquire annual memberships. 

The final product of the project is the recommendations highlighted for increasing annual memberships. I hope that this project serves as a useful example of the application of data analytics principles and methods to solve practical problems.

Thank you for visiting this repository. I welcome any feedback or questions you may have about the project.